Bookmarkable URL
The page now updates the URL to include the following properties as URL
parameters:
- name
- department
- supervisor

The added code checks to see if the above input values are different
than their corresponding URL parameters. If the values do not match, the
page will be redirected to update the URL parameters. At this point, a
cookie will restore the user's hours.
Any time the user submits data, the content is saved into a cookie that
expires in 14 days and restored appropriately. This ensures that no
information is lost if the user navigates away or is redirected
(updating the URL).